Split udp.cpp into per-protocol files, encapsulate two low-fanout globals
udp.cpp mixed at least 6 unrelated wire protocols (WLED's own Notifier
sync, node discovery, Hyperion, TPM2.NET, legacy UDP realtime,
Art-Net/DDP/E1.31 sender, ESP-NOW transport) in one 987-line file,
mostly interleaved inside a single handleNotifications() dispatcher
with nested early-returns. Split each protocol into its own file with
handleNotifications() left as a thin dispatcher:

- sync_notifier.cpp      WLED's own state-sync protocol
- sync_nodes.cpp         node discovery
- realtime.cpp           shared realtime lock/unlock state machine
- realtime_udp.cpp       Hyperion / TPM2.NET / legacy UDP realtime
- realtime_broadcast.cpp outbound Art-Net/DDP/E1.31 sender
- espnow_sync.cpp        ESP-NOW transport for the notifier protocol
- sync.h                 constants shared by the above
- udp.cpp                handleNotifications() dispatcher only

Every extraction preserves the original control flow exactly; bare
`return;` inside the old inline blocks map 1:1 to `return true/false;`
in the new boolean-returning helpers.

Also encapsulated two clusters of WLED_GLOBAL state that turned out to
be used only within this subsystem once split out:
- tpmPacketCount / tpmPayloadFrameSize were referenced only in
  realtime_udp.cpp; now file-local statics instead of globals.
- notificationCount / notificationSentCallMode / notificationSentTime
  were referenced only in udp.cpp + sync_notifier.cpp; now a
  NotifierSendState struct private to sync_notifier.cpp, exposed via
  a single notifyRetryIfNeeded() call instead of raw field access.

Verified: nodemcuv2 (ESP8266) builds and links cleanly, including the
wled_espnow usermod linking against the relocated ESP-NOW callbacks.
esp32dev could not be verified in this environment due to a sandbox
network issue unrelated to this change (opened as draft pending that).
